chore: Enable Materialized View Features for Enterprise Edition
