+++ src/query/ee/src/test_kits/mock_services.rs
@@ -21,6 +21,7 @@ use databend_common_license::license_manager::LicenseManagerSwitch;
 
 use crate::data_mask::RealDatamaskHandler;
 use crate::license::RealLicenseManager;
+use crate::materialized_view::RealMaterializedViewHandler;
 use crate::row_access_policy::row_access_policy_handler::RealRowAccessPolicyHandler;
 use crate::storages::fuse::operations::RealVacuumHandler;
 use crate::stream::RealStreamHandler;
@@ -34,6 +35,7 @@ impl MockServices {
         RealVacuumHandler::init()?;
         RealDatamaskHandler::init()?;
         RealRowAccessPolicyHandler::init()?;
+        RealMaterializedViewHandler::init()?;
         RealStreamHandler::init()?;
         Ok(())
     }
