+++ src/query/service/Cargo.toml
@@ -111,6 +111,7 @@ databend-common-users = { workspace = true }
 databend-common-version = { workspace = true }
 databend-enterprise-attach-table = { workspace = true }
 databend-enterprise-data-mask-feature = { workspace = true }
+databend-enterprise-materialized-view = { workspace = true }
 databend-enterprise-resources-management = { workspace = true }
 databend-enterprise-row-access-policy-feature = { workspace = true }
 databend-enterprise-stream-handler = { workspace = true }
