+++ src/query/service/src/interpreters/interpreter_mutation.rs
@@ -82,7 +82,7 @@ impl MutationInterpreter {
         })
     }
 
-    pub(crate) fn try_create_materialized_view_refresh(
+    pub fn try_create_materialized_view_refresh(
         ctx: Arc<QueryContext>,
         s_expr: SExpr,
         schema: DataSchemaRef,
