+++ src/query/service/src/interpreters/mod.rs
@@ -326,6 +326,7 @@ pub use interpreter_table_undrop::UndropTableInterpreter;
 pub use interpreter_table_vacuum::VacuumTableInterpreter;
 pub use interpreter_tag_create::CreateTagInterpreter;
 pub use interpreter_tag_drop::DropTagInterpreter;
+pub use interpreter_txn_commit::execute_commit_statement;
 pub use interpreter_unset::UnSetInterpreter;
 pub use interpreter_unset_workload_group_quotas::UnsetWorkloadGroupQuotasInterpreter;
 pub use interpreter_use_database::UseDatabaseInterpreter;
